Added PNG optimiser.
The software tries to load PngOptimizerDll at start, then, if the DLL
was found, optimises the PNG images, looks whether they are smaller,
and, if positive, returns the optimised variant.
If  no optimiser present, everything runs as before.
The used DLL was the singleoptimiser that worked with data in memory and
not in a file.