Moved JavaScript contents from index.html to index.js
Added index.js to simplify identation of the JavaScript part. Added many
to to recognise and support input events. Currently work-in-progress on
mouse events.  Mouse down and mouse up are recognised. Mouse moves are
sent as they come, when a mouse button is pressed, and after a 2 seconds
of immobility else. Keyboard input is also here, but currently commented
out.