mousemove event works.
Enabled function sendEvent to send not only event's type, but other
parameters too. Fixed sendEvent calls for mousemove event to include
current mouse position. Adjusted C# part to set mouse position on the
host screen. This makes testing on the same computer a little difficult,
as the cursor position is stolen to the host screen from the guest
screen, after the mousemove event was sent.